Create Ex 6.8 Scanf function
If the data items are entered as
21 10.345 F
When the program is executed, then 21 will be assigned to i , 10.34 will be assigned to x and the character
5 will be assigned to c. The remaining character F will be ignored.

If you specify a field width in scanf(), say %10s, then scanf() collects upto 10 characters or to the first
white space character, whichever comes first. This is true for type int, float and double as well.

The example below demonstrates the use of the scanf() function to enter a string consisting of uppercase
letters and blank spaces. The string will be of undetermined length, but it will be limited to 79 characters
(actually, 80 characters including the null character that is added at the end).