Create Ex 6.10 Scanf function
When the scanf() function is executed, a string of undetermined length (but not more than 79 characters)
will be entered from the standard input device and assigned to line. There will be no restriction on the
characters that compose the string, except that they all fit on one line. For example, the string